initial commit of httpforwarded package
This is the initial commit of the `httpforwarded` package. It contains a
`Format()` and a `Parse()` function. It also tests these functions for
correctness.

Some of the source code contained herein was copied from the Go standard
library. As such, this project retains the copyright of the Go language
and is released under the same license as it is a derivative work.